Inserindo Pedido

+++ src/main/java/com/jeff/api/domain/Produto.java
@@ -24,7 +24,7 @@ public class Produto implements Serializable {
 
 	@Id
 	@GeneratedValue(strategy = GenerationType.IDENTITY)
-	private Long id;
+	private Integer id;
 	private String nome;
 	private Double preco;
 
@@ -41,7 +41,7 @@ public class Produto implements Serializable {
 
 	}
 
-	public Produto(Long id, String nome, Double preco) {
+	public Produto(Integer id, String nome, Double preco) {
 		super();
 		this.id = id;
 		this.nome = nome;
@@ -59,11 +59,11 @@ public class Produto implements Serializable {
 		return lista;
 	}
 
-	public Long getId() {
+	public Integer getId() {
 		return id;
 	}
 
-	public void setId(Long id) {
+	public void setId(Integer id) {
 		this.id = id;
 	}
